+++ flust-engine-sys/build.rs
@@ -37,7 +37,7 @@ fn main() {
 
     let out_path = PathBuf::from(std::env::var("OUT_DIR").unwrap());
     bindings
-        .write_to_file(out_path.join("flutter-engine-sys.rs"))
+        .write_to_file(out_path.join("flust-engine-sys.rs"))
         .expect("Couldn't write bindings!");
 }