+++ flust-engine/src/tasks.rs
@@ -1,6 +1,5 @@
 use crate::FlutterEngineWeakRef;
-use flutter_engine_sys::{FlutterEngineGetCurrentTime, FlutterTask};
-use tracing::debug;
+use flust_engine_sys::{FlutterEngineGetCurrentTime, FlutterTask};
 use parking_lot::{Mutex, MutexGuard};
 use priority_queue::PriorityQueue;
 use std::cmp::Ordering;
@@ -10,6 +9,7 @@ use std::sync::Arc;
 use std::thread;
 use std::thread::ThreadId;
 use std::time::{Duration, Instant};
+use tracing::debug;
 
 pub trait TaskRunnerHandler {
     fn wake(&self);