+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Multi channel plugin support.
 - List scrolling.
 - Update flutter embedder.
-- Add flutter-engine-sys crate to generate bindings using bindgen.
+- Add flust-engine-sys crate to generate bindings using bindgen.
 - flutter-download now download flutter engine of the same version as your flutter sdk.
 - Support linux snap distribution.